third_party/WebKit/LayoutTests/http/tests/notifications/serviceworkerregistration-document-direction.html

<!doctype html>
<html>
  <head>
    <title>Notifications: ServiceWorkerRegistration.showNotification() shows notification with correct notification direction.</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src="../serviceworker/resources/test-helpers.js"></script>
    <script src="resources/test-helpers.js"></script>
  </head>
  <body>
    <script>
      // Tests that the showNotification() function resolves a promise, and that the
      // notificationclick event gets fired with the correct notification direction.
      // This test requires the test runner.

      async_test(function(test) {
          var scope = 'resources/spec/' + location.pathname,
              script = 'resources/instrumentation-service-worker.js';

          testRunner.setPermission('notifications', 'granted', location.origin, location.origin);

          var workerInfo = null;
          getActiveServiceWorkerWithMessagePort(test, script, scope).then(function(info) {
              workerInfo = info;

              // (1) Display a Web Notification from the document.
              return workerInfo.registration.showNotification(scope, {
                  dir: 'rtl'
              });
          }).then(function() {
              // (2) Simulate a click on the notification that has been displayed.
              testRunner.simulateWebNotificationClick(scope, -1 /* action_index */);

              workerInfo.port.addEventListener('message', function(event) {
                  if (typeof event.data != 'object' || !event.data.command) {
                      assert_unreached('Received an invalid message from the Service Worker.');
                      return;
                  }

                  // (3) Verify that the click event was received by the Service Worker,
                  // and that the correct direction was passed along.
                  assert_equals(event.data.command, 'click');
                  assert_equals(event.data.notification.dir, 'rtl');

                  test.done();
              });
          }).catch(unreached_rejection(test));

      }, 'Clicking on a notification displayed through showNotification() fires a Service Worker event with the correct notification direction');
    </script>
  </body>
</html>
